requirements.txt
fastapi==0.135.1
uvicorn==0.42.0

# Database
sqlalchemy==2.0.41
psycopg2-binary==2.9.9

# Validation
pydantic==2.12.5
email-validator==2.3.0

# Async / Server
starlette==0.52.1
anyio==4.12.1
h11==0.16.0

# Env config
python-dotenv==1.0.1

# Redis (for OTP / caching)
redis==7.3.0
fakeredis==2.34.1

# Utils (auto-installed mostly, but safe to keep)
typing_extensions==4.15.0
annotated-types==0.7.0
typing-inspection==0.4.2
idna==3.11
dnspython==2.8.0

# CLI / internal deps
click==8.3.1
colorama==0.4.6

# Optional (good for future migrations)
alembic==1.13.2

celery==5.4.0
PyJWT==2.9.0
bcrypt==4.1.2
razorpay==1.4.2